feat crud Masih Belum Clean Code delete table from component
